Community post and response dto 기초 세팅 with 테스트

+++ packages/server/src/main/java/Doffy/server/community/entity/Comment.java
@@ -17,7 +17,7 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private int commentId;
+    private long commentId;
 
     @Column(name = "comment_body")
     private String commentBody;
